# api/openapi.yaml
openapi: 3.0.3
info:
  title: HydraBridge API
  version: 0.1.0
  summary: Unified control API for HydraRoute Neo
  description: |
    HydraBridge is the control-plane API for HydraRoute Neo. It manages
    configuration files, service lifecycle, backups, status, and logs for
    clients such as iOS apps, web UIs, Telegram bots, CLIs, and automations.
  x-supported-hrneo-version: 3.11.0-1
servers:
  - url: http://{host}/api/v1
    variables:
      host:
        default: 192.168.1.1:2080
security:
  - bearerAuth: []
...
paths:
...
  /status:
    get:
      tags: [Status]
      summary: Return aggregated HydraBridge and hrneo status
      responses:
        "200":
          description: Current status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /compatibility:
    get:
      tags: [Compatibility]
      summary: Check current files against supported HR Neo behavior
      description: |
        Returns a read-only compatibility report for HydraBridge's supported
        HR Neo version. The report validates config syntax, highlights unknown
        or missing keys, checks referenced GeoData files, and surfaces issues
        that would be ignored or warned about by HR Neo at runtime.
      responses:
        "200":
          description: Compatibility report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompatibilityReport"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /doctor:
    get:
      tags: [Doctor]
      summary: Return an actionable router readiness report
      responses:
        "200":
          description: Doctor report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorReport"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /overview:
    get:
      tags: [Overview]
      summary: Return dashboard summary for apps and bots
      responses:
        "200":
          description: Overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverviewResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /auth/token/rotate:
    post:
      tags: [Auth]
      summary: Rotate HydraBridge bearer token
      responses:
        "200":
          description: New token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenRotateResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /config/hrneo/structured:
    get:
      tags: [Config]
      summary: Read hrneo.conf as structured key/value config
      responses:
        "200":
          description: Structured hrneo.conf
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HRNeoConfigResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [Config]
      summary: Merge hrneo.conf from structured key/value config
      description: |
        Writes `hrneo.conf` atomically after creating a backup. Omitted fields
        keep their current values, explicit empty arrays clear repeatable keys,
        and unknown existing keys are preserved for forward compatibility.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutHRNeoConfigRequest"
      responses:
        "200":
          description: Config saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutConfigResponse"
        "202":
          description: Config saved but apply action failed or is incomplete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutConfigResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
  /config/hrneo/default:
    get:
      tags: [Config]
      summary: Preview the default hrneo.conf generated by HR Neo semantics
      responses:
        "200":
          description: Default hrneo.conf preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HRNeoDefaultConfigResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /config/hrneo/generate-default:
    post:
      tags: [Config]
      summary: Replace hrneo.conf with HR Neo defaults
      description: |
        Replaces hrneo.conf with the same 27 default keys produced by
        `hrneo --genconfig`, after creating a backup. Accepts `If-Match`.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateDefaultHRNeoConfigRequest"
      responses:
        "200":
          description: Default config saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutConfigResponse"
        "202":
          description: Default config saved but restart failed or is incomplete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutConfigResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
  /metadata/hrneo-params:
    get:
      tags: [Config]
      summary: Return metadata for all HR Neo configuration parameters
      responses:
        "200":
          description: HR Neo config parameter metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HRNeoParamMetadataResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /config/domains/structured:
    get:
      tags: [Config]
      summary: Read domain.conf as structured targets
      responses:
        "200":
          description: Structured domain.conf
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainConfigResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [Config]
      summary: Replace domain.conf from structured targets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutDomainConfigRequest"
      responses:
        "200":
          description: Config saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutConfigResponse"
        "202":
          description: Config saved but apply action failed or is incomplete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutConfigResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
  /config/domains/validate:
    post:
      tags: [Config]
      summary: Validate domain.conf structure without saving it
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutDomainConfigRequest"
      responses:
        "200":
          description: Domain validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /config/cidr/structured:
    get:
      tags: [Config]
      summary: Read ip.list as structured CIDR blocks
      responses:
        "200":
          description: Structured ip.list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CIDRConfigResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [Config]
      summary: Replace ip.list from structured CIDR blocks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutCIDRConfigRequest"
      responses:
        "200":
          description: Config saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutConfigResponse"
        "202":
          description: Config saved but apply action failed or is incomplete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutConfigResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
  /config/cidr/validate:
    post:
      tags: [Config]
      summary: Validate ip.list structure without saving it
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutCIDRConfigRequest"
      responses:
        "200":
          description: CIDR validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /diagnostics/domain:
    post:
      tags: [Diagnostics]
      summary: Explain how HR Neo would route a domain
      description: |
        Uses HydraBridge's HR Neo-compatible parser and matching rules for
        `domain.conf`: exact match, suffix match, specificity, and PolicyOrder.
        GeoSite `.dat` expansion is reported separately by GeoData APIs and is
        not expanded by this endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DomainDiagnosticRequest"
      responses:
        "200":
          description: Domain diagnostic result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainDiagnosticResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /diagnostics/ip:
    post:
      tags: [Diagnostics]
      summary: Explain how HR Neo would route an IP address
      description: |
        Combines matching static CIDR rules, configured GeoIP `.dat` prefixes,
        live ipset membership, relevant firewall rules, and Keenetic policy
        marks when those runtime sources are available. DNS/CNAME and L7
        history are not fabricated because HR Neo does not expose durable
        attribution for entries already present in ipsets.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IPDiagnosticRequest"
      responses:
        "200":
          description: IP diagnostic result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IPDiagnosticResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /runtime/ipsets:
    get:
      tags: [Runtime]
      summary: Return live ipset names and referenced target set status
      responses:
        "200":
          description: Runtime ipset view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeIPSetsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /runtime/firewall:
    get:
      tags: [Runtime]
      summary: Return live iptables/ip6tables mangle rules relevant to HR Neo
      responses:
        "200":
          description: Runtime firewall view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeFirewallResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /runtime/policies:
    get:
      tags: [Runtime]
      summary: Return live Keenetic policy marks from RCI
      responses:
        "200":
          description: Runtime policy marks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoliciesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /config/domains/targets/{target}/rules:
    parameters:
      - name: target
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Config]
      summary: Add one domain or GeoSite rule to a target
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DomainRulePatchRequest"
      responses:
        "200":
          description: Rule added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutConfigResponse"
        "202":
          description: Rule added but apply action failed or is incomplete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutConfigResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
    delete:
      tags: [Config]
      summary: Remove one domain or GeoSite rule from a target
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DomainRulePatchRequest"
      responses:
        "200":
          description: Rule removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutConfigResponse"
        "202":
          description: Rule removed but apply action failed or is incomplete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutConfigResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
  /config/cidr/targets/{target}/rules:
    parameters:
      - name: target
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Config]
      summary: Add one CIDR or GeoIP rule to a target
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CIDRRulePatchRequest"
      responses:
        "200":
          description: Rule added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutConfigResponse"
        "202":
          description: Rule added but apply action failed or is incomplete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutConfigResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
    delete:
      tags: [Config]
      summary: Remove one CIDR or GeoIP rule from a target
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CIDRRulePatchRequest"
      responses:
        "200":
          description: Rule removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutConfigResponse"
        "202":
          description: Rule removed but apply action failed or is incomplete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutConfigResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
  /views/domains/grouped:
    get:
      tags: [Views]
      summary: Return domain.conf grouped by target for display
      description: |
        Read-only app-facing projection. Repeated target lines are aggregated
        for display without changing the underlying HR Neo config file.
      responses:
        "200":
          description: Grouped domain view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupedDomainViewResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /views/cidr/grouped:
    get:
      tags: [Views]
      summary: Return ip.list grouped by target for display
      description: |
        Read-only app-facing projection. Repeated target blocks are aggregated
        for display without changing the underlying HR Neo config file.
      responses:
        "200":
          description: Grouped CIDR view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupedCIDRViewResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /runtime/direct-routes:
    get:
      tags: [Runtime]
      summary: Return live Linux policy-routing state used by DirectRoute
      responses:
        "200":
          description: Runtime DirectRoute view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeDirectRoutesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /targets:
    get:
      tags: [Targets]
      summary: Return routing targets referenced by HydraRoute configs
      responses:
        "200":
          description: Target inventory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TargetInventoryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /targets/interfaces:
    get:
      tags: [Targets]
      summary: Return system interfaces available for DirectRoute targets
      responses:
        "200":
          description: Interface list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterfacesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /targets/policies:
    get:
      tags: [Targets]
      summary: Return Keenetic policy targets from RCI
      responses:
        "200":
          description: Policy list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoliciesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Targets]
      summary: Create or ensure a Keenetic policy through RCI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyMutationRequest"
      responses:
        "200":
          description: Policy command executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyMutationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /targets/policies/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    delete:
      tags: [Targets]
      summary: Delete a Keenetic policy through RCI
      description: |
        Sends `no ip policy <name>` through Keenetic RCI and saves startup
        configuration. Smoke-test this RCI command on target firmware when
        upgrading beyond the documented HR Neo compatibility version.
      responses:
        "200":
          description: Policy command executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyMutationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /targets/order:
    get:
      tags: [Targets]
      summary: Return configured target priority order
      responses:
        "200":
          description: Target order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TargetOrderResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [Targets]
      summary: Replace PolicyOrder in hrneo.conf
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutTargetOrderRequest"
      responses:
        "200":
          description: Target order saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutConfigResponse"
        "202":
          description: Target order saved but apply action failed or is incomplete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutConfigResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /geodata/files:
    get:
      tags: [GeoData]
      summary: Return configured GeoIP and GeoSite file metadata
      responses:
        "200":
          description: GeoData files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeoDataFilesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [GeoData]
      summary: Upload a GeoIP or GeoSite .dat file from base64 content
      description: The destination path must be inside the geofile directory next to hrneo.conf.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeoDataUploadRequest"
      responses:
        "200":
          description: GeoData file saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeoDataFileMutationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /geodata/download:
    post:
      tags: [GeoData]
      summary: Download a GeoIP or GeoSite .dat file from a URL
      description: The HTTP(S) source must resolve to public IP addresses. Loopback, private, and link-local addresses are rejected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeoDataDownloadRequest"
      responses:
        "200":
          description: GeoData file downloaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeoDataFileMutationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /export/csv:
    get:
      tags: [ImportExport]
      summary: Export domain.conf and ip.list as CSV
      responses:
        "200":
          description: CSV export
          content:
            text/csv:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
...
  responses:
...
    Unauthorized:
      description: Missing bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    StatusResponse:
      type: object
      required: [hrbridge, hrneo, paths]
      properties:
        hrbridge:
          type: object
          required: [version, uptimeSec]
          properties:
            version:
              type: string
            uptimeSec:
              type: integer
              format: int64
        hrneo:
          $ref: "#/components/schemas/HRNeoStatus"
        paths:
          type: object
          required: [hrneoConf, domainConf, cidrList, backupDir, auditLog]
          properties:
            hrneoConf:
              type: string
            domainConf:
              type: string
            cidrList:
              type: string
            backupDir:
              type: string
            auditLog:
              type: string
...